Baseline Web Application
This primary commit includes all of the structural implementation required to have a working environment to run/test/develop on the web app.
Largest voids left in the structure are the completion of the database migrations and the implementation of the email confirmations. The mailer is already implemented, just need the confirmation workflows.